+++ assets/admin/components/slide/slide-form.jsx
@@ -15,7 +15,7 @@ import idFromUrl from "../util/helpers/id-from-url";
 import FormInput from "../util/forms/form-input";
 import ContentForm from "./content/content-form";
 import LoadingComponent from "../util/loading-component/loading-component";
-import SlidePreview from "./preview/slide-preview.jsx";
+import SlidePreview from "./preview/slide-preview";
 import FeedSelector from "./content/feed-selector";
 import SelectPlaylistsTable from "../util/multi-and-table/select-playlists-table";
 import localStorageKeys from "../util/local-storage-keys";
@@ -25,6 +25,7 @@ import "./slide-form.scss";
 import Preview from "../preview/preview";
 import StickyFooter from "../util/sticky-footer";
 import Select from "../util/forms/select";
+import { getConfig } from "../../../shared/slide-utils/templates";
 
 /**
  * The slide form component.
@@ -91,6 +92,7 @@ function SlideForm({
   const [themesOptions, setThemesOptions] = useState();
   const [displayPreview, setDisplayPreview] = useState(null);
   const [templateError, setTemplateError] = useState(false);
+  const [disableLivePreview, setDisableLivePreview] = useState(false);
 
   // Load templates.
   const { data: templates, isLoading: loadingTemplates } =
@@ -156,19 +158,12 @@ function SlideForm({
     const newSelectedTemplates = [];
 
     if (selectedTemplate) {
-      // Get content form from template resources.
-      const contentFormUrl = selectedTemplate?.resources?.admin;
-      fetch(contentFormUrl)
-        .then((response) => response.json())
-        .then((data) => {
-          setContentFormElements(data);
-        })
-        .catch((er) => {
-          displayError(t("template-error"), er);
-        });
-
+      const slideConfig = getConfig(selectedTemplate['id']);
+      setContentFormElements(slideConfig.adminForm ?? []);
+      setDisableLivePreview(slideConfig?.options?.disableLivePreview ?? false);
       newSelectedTemplates.push(selectedTemplate);
     }
+
     setSelectedTemplates(newSelectedTemplates);
   }, [selectedTemplate]);
 
@@ -400,13 +395,12 @@ function SlideForm({
                 </Button>
               </div>
 
-              {selectedTemplate?.resources?.options?.disableLivePreview && (
+              {disableLivePreview && (
                 <Alert variant="secondary" className="mt-3">
                   {t("slide-preview-disabled-preview")}
                 </Alert>
               )}
-              {!selectedTemplate?.resources?.options?.disableLivePreview &&
-                selectedTemplate?.resources?.component && (
+              {!disableLivePreview && (
                   <>
                     {previewOrientation === "horizontal" && (
                       <div style={{ width: "100%" }}>