+++ assets/admin/components/user/login.jsx
@@ -9,13 +9,13 @@ import {MultiSelect} from "react-multi-select-component";
 import UserContext from "../../context/user-context";
 import FormInput from "../util/forms/form-input";
 import { enhancedApi } from "../../../shared/redux/enhanced-api.ts";
-import AdminConfigLoader from "../../admin-config-loader.js";
+import AdminConfigLoader from "../util/admin-config-loader.js";
 import {displayError} from "../util/list/toast-component/display-toast";
 import localStorageKeys from "../util/local-storage-keys";
 import LoginSidebar from "../navigation/login-sidebar/login-sidebar";
-import "./login.scss";
 import OIDCLogin from "./oidc-login";
 import LoadingComponent from "../util/loading-component/loading-component";
+import "./login.scss";
 
 /**
  * Login component