+++ src/Command/UpdateCommand.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Command;
+
+class UpdateCommand
+{
+    // TODO: Test that migrations have been run.
+    // TODO: Run test of status for templates. No templates = clean install. Install all?
+    // TODO: Update existing templates.
+}