+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * Cleaned up Github Actions workflows.
 * Updated PHP dependencies.
 * Added Playwright github action.
+* Changed how templates are imported.
 * Removed propTypes in admin.
 * Upgraded redux-toolkit and how api slices are generated.
 * Fixed redux-toolkit cache handling.